@@ -12,8 +12,8 @@ repositories {
     }
 }
 
-sourceCompatibility = '11'
-targetCompatibility = '11'
+sourceCompatibility = '17'
+targetCompatibility = '17'
 
 configurations {
     compileOnlyA
@@ -49,16 +49,16 @@ test {
 
 task sourcesJar(type: Jar) {
     from sourceSets.main.allJava
-    archiveClassifier = 'sources'
+    archiveClassifier.set('sources')
 }
 
 task javadocJar(type: Jar) {
     from javadoc
-    archiveClassifier = 'javadoc'
+    archiveClassifier.set('javadoc')
 }
 
 task cleanjar(type: Jar, dependsOn: jar) {
-    classifier = 'clean'
+    archiveClassifier.set('clean')
     manifest.attributes("Main-Class": mainClassName,
             "Premain-Class": mainAgentName,
             "Can-Redefine-Classes": "true",
@@ -88,15 +88,15 @@ dependencies {
     bundle group: 'mysql', name: 'mysql-connector-java', version: rootProject['verMySQLConn']
     bundle group: 'org.postgresql', name: 'postgresql', version: rootProject['verPostgreSQLConn']
     bundle group: 'com.guardsquare', name: 'proguard-base', version: rootProject['verProguard']
-    bundle group: 'org.apache.logging.log4j', name: 'log4j-core', version: rootProject['verLog4j']
+    bundle group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.14.1'
     bundle group: 'org.apache.logging.log4j', name: 'log4j-slf4j-impl', version: rootProject['verLog4j']
     bundle group: 'io.jsonwebtoken', name: 'jjwt-api', version: rootProject['verJwt']
     bundle group: 'io.jsonwebtoken', name: 'jjwt-impl', version: rootProject['verJwt']
     bundle group: 'io.jsonwebtoken', name: 'jjwt-gson', version: rootProject['verJwt']
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter', version: rootProject['verJunit']
 
-    hikari 'io.micrometer:micrometer-core:1.5.10'
-    hikari('com.zaxxer:HikariCP:4.0.3') {
+    hikari 'io.micrometer:micrometer-core:1.7.4'
+    hikari('com.zaxxer:HikariCP:5.0.0')  {
         exclude group: 'javassist'
         exclude group: 'io.micrometer'
         exclude group: 'org.slf4j'
@@ -115,7 +115,7 @@ dependencies {
     compileOnlyA group: 'com.google.guava', name: 'guava', version: rootProject['verGuavaC']
     // Do not update (laggy deps).
     compileOnlyA 'log4j:log4j:1.2.17'
-    compileOnlyA 'org.apache.logging.log4j:log4j-core:2.11.2'
+    compileOnlyA 'org.apache.logging.log4j:log4j-core:2.14.1'
 }
 
 task hikari(type: Copy) {
