@@ -4,13 +4,13 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.foxesworld.launcher.Launcher;
 import org.foxesworld.launcher.NeedGarbageCollection;
-import org.foxesworld.launcher.hasher.HashedDir;
 import org.foxesworld.launcher.managers.ConfigManager;
 import org.foxesworld.launcher.managers.GarbageManager;
 import org.foxesworld.launcher.modules.events.ClosePhase;
 import org.foxesworld.launcher.news.News;
 import org.foxesworld.launcher.profiles.ClientProfile;
 import org.foxesworld.launchserver.auth.AuthProviderPair;
+import org.foxesworld.launchserver.auth.core.RejectAuthCoreProvider;
 import org.foxesworld.launchserver.auth.session.MemorySessionStorage;
 import org.foxesworld.launchserver.binary.EXEL4JLauncherBinary;
 import org.foxesworld.launchserver.binary.EXELauncherBinary;
@@ -40,10 +40,7 @@ import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.nio.file.*;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.security.interfaces.ECPrivateKey;
-import java.security.interfaces.ECPublicKey;
 import java.util.*;
-import java.util.Map.Entry;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -89,8 +86,6 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
      * This object contains runtime configuration
      */
     public final LaunchServerRuntimeConfig runtime;
-    public final ECPublicKey publicKey;
-    public final ECPrivateKey privateKey;
     /**
      * Pipeline for building JAR
      */
@@ -119,15 +114,11 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
     // Server
     public final CommandHandler commandHandler;
     public final NettyServerSocketHandler nettyServerSocketHandler;
-    @Deprecated
-    public final Timer taskPool;
     public final ScheduledExecutorService service;
     public final AtomicBoolean started = new AtomicBoolean(false);
     public final LauncherModuleLoader launcherModuleLoader;
     private final Logger logger = LogManager.getLogger();
     public LaunchServerConfig config;
-    @Deprecated
-    public volatile Map<String, HashedDir> updatesDirMap;
     // Updates and profiles
     private volatile Set<ClientProfile> profilesList;
 
@@ -141,13 +132,10 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
         this.profilesDir = directories.profilesDir;
         this.updatesDir = directories.updatesDir;
         this.keyAgreementManager = keyAgreementManager;
-        this.publicKey = keyAgreementManager.ecdsaPublicKey;
-        this.privateKey = keyAgreementManager.ecdsaPrivateKey;
         this.commandHandler = commandHandler;
         this.runtime = runtimeConfig;
         this.certificateManager = certificateManager;
         this.service = Executors.newScheduledThreadPool(config.netty.performance.schedulerThread);
-        taskPool = new Timer("Timered task worker thread", true);
         launcherLibraries = directories.launcherLibrariesDir;
         launcherLibrariesCompile = directories.launcherLibrariesCompileDir;
         config.setLaunchServer(this);
@@ -237,7 +225,7 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
     @Override
     public Map<String, Command> getCommands() {
         Map<String, Command> commands = new HashMap<>();
-        SubCommand reload = new SubCommand() {
+        SubCommand reload = new SubCommand("[type]", "reload launchserver config") {
             @Override
             public void invoke(String... args) throws Exception {
                 if (args.length == 0) {
@@ -245,20 +233,14 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
                     return;
                 }
                 switch (args[0]) {
-                    case "no_auth":
-                        reload(ReloadType.NO_AUTH);
-                        break;
-                    case "no_components":
-                        reload(ReloadType.NO_COMPONENTS);
-                        break;
-                    default:
-                        reload(ReloadType.FULL);
-                        break;
+                    case "no_auth" -> reload(ReloadType.NO_AUTH);
+                    case "no_components" -> reload(ReloadType.NO_COMPONENTS);
+                    default -> reload(ReloadType.FULL);
                 }
             }
         };
         commands.put("reload", reload);
-        SubCommand save = new SubCommand() {
+        SubCommand save = new SubCommand("[]", "save launchserver config") {
             @Override
             public void invoke(String... args) throws Exception {
                 launchServerConfigManager.writeConfig(config);
@@ -267,6 +249,22 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
             }
         };
         commands.put("save", save);
+        LaunchServer instance = this;
+        SubCommand resetauth = new SubCommand("authId", "reset auth by id") {
+            @Override
+            public void invoke(String... args) throws Exception {
+                verifyArgs(args, 1);
+                AuthProviderPair pair = config.getAuthProviderPair(args[0]);
+                if (pair == null) {
+                    logger.error("Pair not found");
+                    return;
+                }
+                pair.core.close();
+                pair.core = new RejectAuthCoreProvider();
+                pair.core.init(instance);
+            }
+        };
+        commands.put("resetauth", resetauth);
         return commands;
     }
 
@@ -311,16 +309,6 @@ public final class LaunchServer implements Runnable, AutoCloseable, Reconfigurab
 
     public void setProfiles(Set<ClientProfile> profilesList) {
         this.profilesList = Collections.unmodifiableSet(profilesList);
-    }
-
-    @Deprecated
-    public HashedDir getUpdateDir(String name) {
-        return updatesDirMap.get(name);
-    }
-
-    @Deprecated
-    public Set<Entry<String, HashedDir>> getUpdateDirs() {
-        return updatesDirMap.entrySet();
     }
 
     public void rebindNettyServerSocket() {
