@@ -1,5 +1,7 @@
 package org.foxesworld.launchserver.manangers;
 
+import io.jsonwebtoken.JwtParser;
+import io.jsonwebtoken.Jwts;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.foxesworld.launcher.ClientPermissions;
@@ -16,12 +18,10 @@ import org.foxesworld.launchserver.auth.core.AuthSocialProvider;
 import org.foxesworld.launchserver.auth.core.User;
 import org.foxesworld.launchserver.auth.core.UserSession;
 import org.foxesworld.launchserver.auth.core.interfaces.user.UserSupportTextures;
-import org.foxesworld.launchserver.auth.provider.AuthProvider;
-import org.foxesworld.launchserver.auth.provider.AuthProviderDAOResult;
-import org.foxesworld.launchserver.auth.provider.AuthProviderResult;
 import org.foxesworld.launchserver.auth.texture.TextureProvider;
 import org.foxesworld.launchserver.socket.Client;
 import org.foxesworld.launchserver.socket.response.auth.AuthResponse;
+import org.foxesworld.launchserver.socket.response.auth.RestoreResponse;
 import org.foxesworld.utils.helper.IOHelper;
 import org.foxesworld.utils.helper.SecurityHelper;
 
@@ -36,9 +36,58 @@ import java.util.stream.Collectors;
 public class AuthManager {
     private transient final LaunchServer server;
     private transient final Logger logger = LogManager.getLogger();
+    private transient final JwtParser checkServerTokenParser;
 
     public AuthManager(LaunchServer server) {
         this.server = server;
+        this.checkServerTokenParser = Jwts.parserBuilder()
+                .requireIssuer("LaunchServer")
+                .require("tokenType", "checkServer")
+                .setSigningKey(server.keyAgreementManager.ecdsaPublicKey)
+                .build();
+    }
+
+    public String newCheckServerToken(String serverName, String authId) {
+        return Jwts.builder()
+                .setIssuer("LaunchServer")
+                .claim("serverName", serverName)
+                .claim("authId", authId)
+                .claim("tokenType", "checkServer")
+                .signWith(server.keyAgreementManager.ecdsaPrivateKey)
+                .compact();
+    }
+
+    public record CheckServerTokenInfo(String serverName, String authId) {
+    }
+
+    public CheckServerTokenInfo parseCheckServerToken(String token) {
+        try {
+            var jwt = checkServerTokenParser.parseClaimsJws(token).getBody();
+            return new CheckServerTokenInfo(jwt.get("serverName", String.class), jwt.get("authId", String.class));
+        } catch (Exception e) {
+            return null;
+        }
+    }
+
+    public static class CheckServerVerifier implements RestoreResponse.ExtendedTokenProvider {
+        private final LaunchServer server;
+
+        public CheckServerVerifier(LaunchServer server) {
+            this.server = server;
+        }
+
+        @Override
+        public boolean accept(Client client, AuthProviderPair pair, String extendedToken) {
+            var info = server.authManager.parseCheckServerToken(extendedToken);
+            if(info == null) {
+                return false;
+            }
+            client.auth_id = info.authId;
+            client.auth = server.config.getAuthProviderPair(info.authId);
+            if(client.permissions == null) client.permissions = new ClientPermissions();
+            client.permissions.addAction("launchserver\\.checkserver");
+            return true;
+        }
     }
 
     /**
@@ -61,11 +110,10 @@ public class AuthManager {
      */
     public void check(AuthResponse.AuthContext context) throws AuthException {
         if (context.authType == AuthResponse.ConnectTypes.CLIENT && !context.client.checkSign) {
-            AuthProvider.authError("Don't skip Launcher Update");
-            return;
+            throw new AuthException("Don't skip Launcher Update");
         }
         if (context.client.isAuth) {
-            AuthProvider.authError("You are already logged in");
+            throw new AuthException("You are already logged in");
         }
     }
 
@@ -77,47 +125,9 @@ public class AuthManager {
      * @return Access token
      */
     public AuthReport auth(AuthResponse.AuthContext context, AuthRequest.AuthPasswordInterface password) throws AuthException {
-        AuthProviderPair pair = context.pair;
-        AuthReport report;
-        if (pair.core == null) {
-            try {
-                report = AuthReport.ofMinecraftAccessToken(authWithProviderAndHandler(context, password));
-            } catch (Exception e) {
-                if (e instanceof AuthException) throw (AuthException) e;
-                throw new AuthException("Internal Auth Error. Please contact administrator ");
-            }
-        } else {
-            report = authWithCore(context, password);
-        }
-        return report;
-    }
-
-    @SuppressWarnings("deprecation")
-    private String authWithProviderAndHandler(AuthResponse.AuthContext context, AuthRequest.AuthPasswordInterface password) throws Exception {
-        String accessToken;
-        context.pair.provider.preAuth(context.login, password, context.ip);
-        AuthProviderResult aresult = context.pair.provider.auth(context.login, password, context.ip, context.hwid);
-        UUID uuid;
-        String username = aresult.username != null ? aresult.username : context.login;
-        if (aresult instanceof AuthProviderDAOResult) {
-            context.client.daoObject = ((AuthProviderDAOResult) aresult).daoObject;
-        }
-        if (context.authType == AuthResponse.ConnectTypes.CLIENT && server.config.protectHandler.allowGetAccessToken(context)) {
-            uuid = context.pair.handler.auth(aresult);
-            accessToken = aresult.accessToken;
-        } else {
-            uuid = context.pair.handler.usernameToUUID(aresult.username);
-            accessToken = null;
-        }
-        internalAuth(context.client, context.authType, context.pair, username, aresult.groupId, uuid, aresult.balance, aresult.permissions, false);
-        return accessToken;
-    }
-
-    private AuthReport authWithCore(AuthResponse.AuthContext context, AuthRequest.AuthPasswordInterface password) throws AuthException {
         AuthCoreProvider provider = context.pair.core;
         provider.verifyAuth(context);
-        if (password instanceof AuthOAuthPassword) {
-            AuthOAuthPassword password1 = (AuthOAuthPassword) password;
+        if (password instanceof AuthOAuthPassword password1) {
             UserSession session;
             try {
                 session = provider.getUserSessionByOAuthAccessToken(password1.accessToken);
@@ -206,9 +216,7 @@ public class AuthManager {
         client.useOAuth = oauth;
         client.balance = balance;
         client.groupId = groupId;
-        if (pair.isUseCore() && client.coreObject == null) {
-            client.coreObject = pair.core.getUserByUUID(uuid);
-        }
+        client.coreObject = pair.core.getUserByUUID(uuid);
     }
 
     public static class CheckServerReport {
@@ -233,23 +241,14 @@ public class AuthManager {
 
     public CheckServerReport checkServer(Client client, String username, String serverID) throws IOException {
         if (client.auth == null) return null;
-        if (client.auth.isUseCore()) {
-            User user = client.auth.core.checkServer(client, username, serverID);
-            if (user == null) return null;
-            else return CheckServerReport.ofUser(user, getPlayerProfile(client.auth, user));
-        } else {
-            UUID uuid = client.auth.handler.checkServer(username, serverID);
-            return uuid == null ? null : CheckServerReport.ofUUID(uuid, getPlayerProfile(client.auth, uuid));
-        }
+        User user = client.auth.core.checkServer(client, username, serverID);
+        if (user == null) return null;
+        else return CheckServerReport.ofUser(user, getPlayerProfile(client.auth, user));
     }
 
     public boolean joinServer(Client client, String username, String accessToken, String serverID) throws IOException {
         if (client.auth == null) return false;
-        if (client.auth.isUseCore()) {
-            return client.auth.core.joinServer(client, username, accessToken, serverID);
-        } else {
-            return client.auth.handler.joinServer(username, accessToken, serverID);
-        }
+        return client.auth.core.joinServer(client, username, accessToken, serverID);
     }
 
     public PlayerProfile getPlayerProfile(Client client) {
@@ -257,6 +256,9 @@ public class AuthManager {
         PlayerProfile playerProfile;
         if (client.useOAuth) {
             User user = client.getUser();
+            if (user == null) {
+                return null;
+            }
             playerProfile = getPlayerProfile(client.auth, user);
             if (playerProfile != null) return playerProfile;
         }
@@ -272,19 +274,14 @@ public class AuthManager {
     }
 
     public PlayerProfile getPlayerProfile(AuthProviderPair pair, String username, ClientProfile profile) {
-        UUID uuid = null;
-        if (pair.isUseCore()) {
-            User user = pair.core.getUserByUsername(username);
-            PlayerProfile playerProfile = getPlayerProfile(pair, user);
-            uuid = user.getUUID();
-            if (playerProfile != null) return playerProfile;
-        } else {
-            try {
-                uuid = pair.handler.usernameToUUID(username);
-            } catch (IOException e) {
-                logger.error("UsernameToUUID failed", e);
-            }
+        UUID uuid;
+        User user = pair.core.getUserByUsername(username);
+        if (user == null) {
+            return null;
         }
+        PlayerProfile playerProfile = getPlayerProfile(pair, user);
+        uuid = user.getUUID();
+        if (playerProfile != null) return playerProfile;
         if (uuid == null) {
             return null;
         }
@@ -299,19 +296,14 @@ public class AuthManager {
     }
 
     public PlayerProfile getPlayerProfile(AuthProviderPair pair, UUID uuid, ClientProfile profile) {
-        String username = null;
-        if (pair.isUseCore()) {
-            User user = pair.core.getUserByUUID(uuid);
-            PlayerProfile playerProfile = getPlayerProfile(pair, user);
-            username = user.getUsername();
-            if (playerProfile != null) return playerProfile;
-        } else {
-            try {
-                username = pair.handler.uuidToUsername(uuid);
-            } catch (IOException e) {
-                logger.error("UUIDToUsername failed", e);
-            }
+        String username;
+        User user = pair.core.getUserByUUID(uuid);
+        if (user == null) {
+            return null;
         }
+        PlayerProfile playerProfile = getPlayerProfile(pair, user);
+        username = user.getUsername();
+        if (playerProfile != null) return playerProfile;
         if (username == null) {
             return null;
         }
@@ -325,6 +317,9 @@ public class AuthManager {
         if (user instanceof UserSupportTextures) {
             return new PlayerProfile(user.getUUID(), user.getUsername(), ((UserSupportTextures) user).getSkinTexture(), ((UserSupportTextures) user).getCloakTexture());
         }
+        if (pair.textureProvider == null) {
+            throw new NullPointerException("TextureProvider not found");
+        }
         return getPlayerProfile(user.getUUID(), user.getUsername(), "", pair.textureProvider);
     }
 
@@ -337,12 +332,10 @@ public class AuthManager {
     }
 
     public AuthRequest.AuthPasswordInterface decryptPassword(AuthRequest.AuthPasswordInterface password) throws AuthException {
-        if (password instanceof Auth2FAPassword) {
-            Auth2FAPassword auth2FAPassword = (Auth2FAPassword) password;
+        if (password instanceof Auth2FAPassword auth2FAPassword) {
             auth2FAPassword.firstPassword = tryDecryptPasswordPlain(auth2FAPassword.firstPassword);
             auth2FAPassword.secondPassword = tryDecryptPasswordPlain(auth2FAPassword.secondPassword);
-        } else if (password instanceof AuthMultiPassword) {
-            AuthMultiPassword multiPassword = (AuthMultiPassword) password;
+        } else if (password instanceof AuthMultiPassword multiPassword) {
             List<AuthRequest.AuthPasswordInterface> list = new ArrayList<>(multiPassword.list.size());
             for (AuthRequest.AuthPasswordInterface p : multiPassword.list) {
                 list.add(tryDecryptPasswordPlain(p));
@@ -354,16 +347,7 @@ public class AuthManager {
         return password;
     }
 
-    @SuppressWarnings("deprecation")
     private AuthRequest.AuthPasswordInterface tryDecryptPasswordPlain(AuthRequest.AuthPasswordInterface password) throws AuthException {
-        if (password instanceof AuthECPassword) {
-            try {
-                return new AuthPlainPassword(IOHelper.decode(SecurityHelper.decrypt(server.runtime.passwordEncryptKey
-                        , ((AuthECPassword) password).password)));
-            } catch (Exception ignored) {
-                throw new AuthException("Password decryption error");
-            }
-        }
         if (password instanceof AuthAESPassword) {
             try {
                 return new AuthPlainPassword(IOHelper.decode(SecurityHelper.decrypt(server.runtime.passwordEncryptKey
@@ -385,20 +369,9 @@ public class AuthManager {
         return password;
     }
 
-    public static class AuthReport {
-        public final String minecraftAccessToken;
-        public final String oauthAccessToken;
-        public final String oauthRefreshToken;
-        public final long oauthExpire;
-        public final UserSession session;
-
-        public AuthReport(String minecraftAccessToken, String oauthAccessToken, String oauthRefreshToken, long oauthExpire, UserSession session) {
-            this.minecraftAccessToken = minecraftAccessToken;
-            this.oauthAccessToken = oauthAccessToken;
-            this.oauthRefreshToken = oauthRefreshToken;
-            this.oauthExpire = oauthExpire;
-            this.session = session;
-        }
+    public record AuthReport(String minecraftAccessToken, String oauthAccessToken,
+                             String oauthRefreshToken, long oauthExpire,
+                             UserSession session) {
 
         public static AuthReport ofOAuth(String oauthAccessToken, String oauthRefreshToken, long oauthExpire) {
             return new AuthReport(null, oauthAccessToken, oauthRefreshToken, oauthExpire, null);
