@@ -128,9 +128,10 @@ public class ServerMenuScene extends AbstractScene {
                         try {
                             ServerPinger pinger = new ServerPinger(serverProfile, profile.getVersion());
                             ServerPinger.Result result = pinger.ping();
-                            application.stateService.addServerSocketPing(serverProfile, result);
-                        } catch (IOException e) {
-                            LogHelper.error(e);
+                            contextHelper.runInFxThread(() -> {
+                                application.stateService.addServerSocketPing(serverProfile, result);
+                            });
+                        } catch (IOException ignored) {
                         }
                     }
                 }).start();
