@@ -1,12 +1,9 @@
 package org.foxesworld.launcher.client.gui.scenes.update;
 
-import javafx.application.Platform;
 import javafx.beans.property.DoubleProperty;
 import javafx.scene.control.*;
-import javafx.scene.shape.Circle;
 import javafx.scene.text.Text;
 import org.foxesworld.launcher.AsyncDownloader;
-import org.foxesworld.launcher.LauncherEngine;
 import org.foxesworld.launcher.client.gui.JavaFXApplication;
 import org.foxesworld.launcher.client.gui.helper.LookupHelper;
 import org.foxesworld.launcher.client.gui.impl.ContextHelper;
@@ -27,7 +24,8 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.*;
-import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Consumer;
 
@@ -182,6 +180,7 @@ public class UpdateScene extends AbstractScene {
                     });
                     LogHelper.info("Diff %d %d", diff.mismatch.size(), diff.extra.size());
                     ContextHelper.runInFxThreadStatic(() -> addLog(String.format("Downloading %s...", dirName)));
+                    ExecutorService executor = Executors.newWorkStealingPool(4);
                     downloader = Downloader.downloadList(adds, updateRequestEvent.url, dir, new Downloader.DownloadCallback() {
                         @Override
                         public void apply(long fullDiff) {
@@ -195,7 +194,7 @@ public class UpdateScene extends AbstractScene {
                         public void onComplete(Path path) {
 
                         }
-                    }, null, 4);
+                    }, executor, 4);
                     downloader.getFuture().thenAccept((e) -> {
                         ContextHelper.runInFxThreadStatic(() -> addLog(String.format("Delete Extra files %s", dirName)));
                         try {
@@ -207,8 +206,9 @@ public class UpdateScene extends AbstractScene {
                     }).exceptionally((e) -> {
                         ContextHelper.runInFxThreadStatic(() -> errorHandle(e));
                         return null;
+                    }).thenAccept((e) -> {
+                        executor.shutdown();
                     });
-
                 } catch (Exception e) {
                     ContextHelper.runInFxThreadStatic(() -> errorHandle(e));
                 }
