@@ -4,109 +4,31 @@ package org.foxesworld.launchserver.manangers;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
-import org.bouncycastle.asn1.x500.X500Name;
-import org.bouncycastle.asn1.x500.X500NameBuilder;
-import org.bouncycastle.asn1.x500.style.BCStyle;
-import org.bouncycastle.asn1.x509.AlgorithmIdentifier;
-import org.bouncycastle.asn1.x509.SubjectPublicKeyInfo;
 import org.bouncycastle.cert.X509CertificateHolder;
-import org.bouncycastle.cert.X509v3CertificateBuilder;
 import org.bouncycastle.crypto.params.AsymmetricKeyParameter;
 import org.bouncycastle.crypto.util.PrivateKeyFactory;
 import org.bouncycastle.crypto.util.PrivateKeyInfoFactory;
-import org.bouncycastle.operator.ContentSigner;
-import org.bouncycastle.operator.DefaultDigestAlgorithmIdentifierFinder;
-import org.bouncycastle.operator.OperatorCreationException;
-import org.bouncycastle.operator.bc.BcECContentSignerBuilder;
-import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 import org.bouncycastle.util.io.pem.PemObject;
 import org.bouncycastle.util.io.pem.PemReader;
 import org.bouncycastle.util.io.pem.PemWriter;
 import org.foxesworld.launcher.LauncherTrustManager;
 import org.foxesworld.utils.helper.IOHelper;
 import org.foxesworld.utils.helper.JVMHelper;
-import org.foxesworld.utils.helper.SecurityHelper;
 
 import java.io.*;
-import java.math.BigInteger;
 import java.net.URL;
 import java.nio.file.*;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.security.*;
+import java.security.PrivateKey;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
-import java.security.spec.ECGenParameterSpec;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 public class CertificateManager {
-    public final int validDays = 60;
-    public final int minusHours = 6;
     private transient final Logger logger = LogManager.getLogger();
-    public X509CertificateHolder ca;
-    public AsymmetricKeyParameter caKey;
-    public X509CertificateHolder server;
-    @SuppressWarnings("unused")
-    public AsymmetricKeyParameter serverKey;
     public LauncherTrustManager trustManager;
-    public String orgName;
-
-    public X509CertificateHolder generateCertificate(String subjectName, PublicKey subjectPublicKey) throws OperatorCreationException {
-        SubjectPublicKeyInfo subjectPubKeyInfo = SubjectPublicKeyInfo.getInstance(subjectPublicKey.getEncoded());
-        BigInteger serial = BigInteger.valueOf(SecurityHelper.newRandom().nextLong());
-        Date startDate = Date.from(Instant.now().minus(minusHours, ChronoUnit.HOURS));
-        Date endDate = Date.from(startDate.toInstant().plus(validDays, ChronoUnit.DAYS));
-
-        X500NameBuilder subject = new X500NameBuilder();
-        subject.addRDN(BCStyle.CN, subjectName);
-        subject.addRDN(BCStyle.O, orgName);
-        X509v3CertificateBuilder v3CertGen = new X509v3CertificateBuilder(ca.getSubject(), serial,
-                startDate, endDate, subject.build(), subjectPubKeyInfo);
-
-        AlgorithmIdentifier sigAlgId = ca.getSignatureAlgorithm();
-        AlgorithmIdentifier digAlgId = new DefaultDigestAlgorithmIdentifierFinder().find(sigAlgId);
-        ContentSigner sigGen = new BcECContentSignerBuilder(sigAlgId, digAlgId).build(caKey);
-
-        return v3CertGen.build(sigGen);
-    }
-
-    public void generateCA() throws NoSuchAlgorithmException, IOException, OperatorCreationException, InvalidAlgorithmParameterException {
-        ECGenParameterSpec ecGenSpec = new ECGenParameterSpec("secp384k1");
-        KeyPairGenerator generator = KeyPairGenerator.getInstance("EC");
-        generator.initialize(ecGenSpec, SecurityHelper.newRandom());
-        KeyPair pair = generator.generateKeyPair();
-        LocalDateTime startDate = LocalDate.now().atStartOfDay();
-
-        X500NameBuilder subject = new X500NameBuilder();
-        subject.addRDN(BCStyle.CN, orgName.concat(" CA"));
-        subject.addRDN(BCStyle.O, orgName);
-
-        X509v3CertificateBuilder builder = new X509v3CertificateBuilder(
-                subject.build(),
-                new BigInteger("0"),
-                Date.from(startDate.atZone(ZoneId.systemDefault()).toInstant()),
-                Date.from(startDate.plusDays(3650).atZone(ZoneId.systemDefault()).toInstant()),
-                new X500Name("CN=ca"),
-                SubjectPublicKeyInfo.getInstance(pair.getPublic().getEncoded()));
-        JcaContentSignerBuilder csBuilder = new JcaContentSignerBuilder("SHA256WITHECDSA");
-        ContentSigner signer = csBuilder.build(pair.getPrivate());
-        ca = builder.build(signer);
-        caKey = PrivateKeyFactory.createKey(pair.getPrivate().getEncoded());
-    }
-
-    public KeyPair generateKeyPair() throws InvalidAlgorithmParameterException, NoSuchAlgorithmException {
-        ECGenParameterSpec ecGenSpec = new ECGenParameterSpec("secp384k1");
-        KeyPairGenerator generator = KeyPairGenerator.getInstance("EC");
-        generator.initialize(ecGenSpec, SecurityHelper.newRandom());
-        return generator.generateKeyPair();
-    }
 
     public void writePrivateKey(Path file, PrivateKey privateKey) throws IOException {
         writePrivateKey(IOHelper.newWriter(file), privateKey);
@@ -203,24 +125,6 @@ public class CertificateManager {
             }
         }, false);
         trustManager = new LauncherTrustManager(certificates.toArray(new X509Certificate[0]));
-    }
-
-    @Deprecated
-    public void checkClass(Class<?> clazz, LauncherTrustManager.CheckMode mode) throws SecurityException {
-        if (trustManager == null) return;
-        X509Certificate[] certificates = JVMHelper.getCertificates(clazz);
-        if (certificates == null) {
-            if (mode == LauncherTrustManager.CheckMode.EXCEPTION_IN_NOT_SIGNED)
-                throw new SecurityException(String.format("Class %s not signed", clazz.getName()));
-            else if (mode == LauncherTrustManager.CheckMode.WARN_IN_NOT_SIGNED)
-                logger.warn("Class {} not signed", clazz.getName());
-            return;
-        }
-        try {
-            trustManager.checkCertificatesSuccess(certificates, trustManager::stdCertificateChecker);
-        } catch (Exception e) {
-            throw new SecurityException(e);
-        }
     }
 
     public LauncherTrustManager.CheckClassResult checkClass(Class<?> clazz) {
