@@ -11,7 +11,7 @@ import org.foxesworld.launcher.Launcher;
 import org.foxesworld.launchserver.socket.NettyConnectContext;
 
 import java.net.URLDecoder;
-import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
@@ -82,7 +82,7 @@ public class NettyWebAPIHandler extends SimpleChannelInboundHandler<FullHttpRequ
             String[] result = sub.split("&");
             Map<String, String> map = new HashMap<>();
             for (String s : result) {
-                String c = URLDecoder.decode(s, Charset.defaultCharset());
+                String c = URLDecoder.decode(s, StandardCharsets.UTF_8);
                 int index = c.indexOf("=");
                 if (index <= 0) {
                     continue;
