@@ -5,6 +5,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.foxesworld.launcher.events.request.CheckServerRequestEvent;
 import org.foxesworld.launchserver.auth.AuthException;
+import org.foxesworld.launchserver.auth.AuthProviderPair;
 import org.foxesworld.launchserver.manangers.AuthManager;
 import org.foxesworld.launchserver.socket.Client;
 import org.foxesworld.launchserver.socket.response.SimpleResponse;
@@ -23,7 +24,7 @@ public class CheckServerResponse extends SimpleResponse {
 
     @Override
     public void execute(ChannelHandlerContext ctx, Client pClient) {
-        if (!pClient.isAuth || pClient.type == AuthResponse.ConnectTypes.CLIENT) {
+        if (pClient.permissions == null || !pClient.permissions.hasAction("launchserver.checkserver")) {
             sendError("Permissions denied");
             return;
         }
