@@ -10,6 +10,7 @@ import org.foxesworld.launchserver.auth.core.AuthCoreProvider;
 import org.foxesworld.launchserver.auth.core.User;
 import org.foxesworld.launchserver.auth.core.UserSession;
 import org.foxesworld.launchserver.auth.protect.AdvancedProtectHandler;
+import org.foxesworld.launchserver.manangers.AuthManager;
 import org.foxesworld.launchserver.socket.Client;
 import org.foxesworld.launchserver.socket.response.SimpleResponse;
 import org.foxesworld.launchserver.socket.response.update.LauncherResponse;
@@ -32,6 +33,7 @@ public class RestoreResponse extends SimpleResponse {
             providers.put(LauncherRequestEvent.LAUNCHER_EXTENDED_TOKEN_NAME, new LauncherResponse.LauncherTokenVerifier(server));
             providers.put("publicKey", new AdvancedProtectHandler.PublicKeyTokenVerifier(server));
             providers.put("hardware", new AdvancedProtectHandler.HardwareInfoTokenVerifier(server));
+            providers.put("checkServer", new AuthManager.CheckServerVerifier(server));
             registeredProviders = true;
         }
     }
@@ -57,7 +59,7 @@ public class RestoreResponse extends SimpleResponse {
         } else {
             pair = client.auth;
         }
-        if (pair == null || !pair.isUseCore()) {
+        if (pair == null) {
             sendError("Invalid authId");
             return;
         }
