@@ -1,9 +1,12 @@
 package org.foxesworld.launcher;
 
 import org.foxesworld.launcher.serialize.HInput;
-import org.foxesworld.launcher.serialize.HOutput;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.regex.Pattern;
 
 public class ClientPermissions {
     public static final ClientPermissions DEFAULT = new ClientPermissions();
@@ -11,6 +14,12 @@ public class ClientPermissions {
     public long permissions;
     @LauncherNetworkAPI
     public long flags;
+    @LauncherNetworkAPI
+    private List<String> roles;
+    @LauncherNetworkAPI
+    private List<String> actions;
+
+    private transient List<Pattern> available;
 
     public ClientPermissions(HInput input) throws IOException {
         this(input.readLong());
@@ -32,51 +41,113 @@ public class ClientPermissions {
     public static ClientPermissions getSuperuserAccount() {
         ClientPermissions perm = new ClientPermissions();
         perm.setPermission(PermissionConsts.ADMIN, true);
+        perm.addAction("*");
         return perm;
+    }
+
+    public boolean hasRole(String role) {
+        return roles != null && roles.contains(role);
+    }
+
+    public synchronized void compile() {
+        if (available != null) {
+            return;
+        }
+        available = new ArrayList<>(actions.size());
+        for (String a : actions) {
+            available.add(Pattern.compile(a));
+        }
+        if (permissions != 0) {
+            if (isPermission(PermissionConsts.ADMIN)) {
+                roles.add("ADMIN");
+                available.add(Pattern.compile(".*"));
+            }
+        }
+    }
+
+    public boolean hasAction(String action) {
+        if (available == null) {
+            compile();
+        }
+        for (Pattern p : available) {
+            if (p.matcher(action).matches()) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public void addRole(String role) {
+        if (roles == null) {
+            roles = new ArrayList<>(1);
+        }
+        roles.add(role);
+    }
+
+    public void addAction(String action) {
+        if (actions == null) {
+            actions = new ArrayList<>(1);
+        }
+        actions.add(action);
+        if(available == null) {
+            available = new ArrayList<>(1);
+        }
+        available.add(Pattern.compile(action));
+    }
+
+    public List<String> getRoles() {
+        return roles;
+    }
+
+    public List<String> getActions() {
+        return actions;
     }
 
     public long toLong() {
         return permissions;
     }
 
-    @Deprecated
-    public void write(HOutput output) throws IOException {
-        output.writeLong(toLong());
-    }
-
     //Read methods
+    @Deprecated
     public final boolean isPermission(PermissionConsts con) {
         return (permissions & con.mask) != 0;
     }
 
+    @Deprecated
     public final boolean isPermission(long mask) {
         return (permissions & mask) != 0;
     }
 
+    @Deprecated
     public final boolean isFlag(FlagConsts con) {
         return (flags & con.mask) != 0;
     }
 
+    @Deprecated
     public final boolean isFlag(long mask) {
         return (flags & mask) != 0;
     }
 
     //Write methods
+    @Deprecated
     public final void setPermission(PermissionConsts con, boolean value) {
         if (value) this.permissions |= con.mask;
         else this.permissions &= ~con.mask;
     }
 
+    @Deprecated
     public final void setPermission(long mask, boolean value) {
         if (value) this.permissions |= mask;
         else this.permissions &= ~mask;
     }
 
+    @Deprecated
     public final void setFlag(FlagConsts con, boolean value) {
         if (value) this.flags |= con.mask;
         else this.flags &= ~con.mask;
     }
 
+    @Deprecated
     public final void setFlag(long mask, boolean value) {
         if (value) this.flags |= mask;
         else this.flags &= ~mask;
@@ -85,11 +156,12 @@ public class ClientPermissions {
     @Override
     public String toString() {
         return "ClientPermissions{" +
-                "permissions=" + permissions +
-                ", flags=" + flags +
+                "roles=" + String.join(", ", roles == null ? Collections.emptyList() : roles) +
+                ", actions=" + String.join(", ", actions == null ? Collections.emptyList() : actions) +
                 '}';
     }
 
+    @Deprecated
     public enum PermissionConsts {
         ADMIN(0x01),
         MANAGEMENT(0x02);
@@ -100,7 +172,7 @@ public class ClientPermissions {
         }
     }
 
-
+    @Deprecated
     public enum FlagConsts {
         SYSTEM(0x01),
         BANNED(0x02),
