@@ -12,9 +12,9 @@ public class AuthRequestEvent extends RequestEvent {
     public static final String ONE_FACTOR_NEED_ERROR_MESSAGE_PREFIX = "auth.require.factor.";
     public static final String OAUTH_TOKEN_EXPIRE = "auth.expiretoken";
     public static final String OAUTH_TOKEN_INVALID = "auth.invalidtoken";
-    public static final String USER_NOT_FOUND_ERROR_MESSAGE = "auth.message.usernotfound";
-    public static final String WRONG_PASSWORD_ERROR_MESSAGE = "auth.message.wrongpassword";
-    public static final String ACCOUNT_BLOCKED_ERROR_MESSAGE = "auth.message.blocked";
+    public static final String USER_NOT_FOUND_ERROR_MESSAGE = "auth.usernotfound";
+    public static final String WRONG_PASSWORD_ERROR_MESSAGE = "auth.wrongpassword";
+    public static final String ACCOUNT_BLOCKED_ERROR_MESSAGE = "auth.userblocked";
     @LauncherNetworkAPI
     public ClientPermissions permissions;
     @LauncherNetworkAPI
